Move hyperopt to separate `BacktestNode` type
