+++ nautilus_trader/backtest/node.py
@@ -15,7 +15,6 @@
 
 import itertools
 import pickle
-from decimal import Decimal
 from typing import Dict, List, Optional
 
 import cloudpickle
@@ -25,13 +24,6 @@ from dask.base import normalize_token
 from dask.delayed import Delayed
 from dask.utils import parse_timedelta
 
-
-try:
-    import hyperopt
-except ImportError:
-    # hyperopt is an optional extra, which is only required when running `hyperopt_search()`.
-    hyperopt = None
-
 from nautilus_trader.backtest.config import BacktestDataConfig
 from nautilus_trader.backtest.config import BacktestRunConfig
 from nautilus_trader.backtest.config import BacktestVenueConfig
@@ -39,12 +31,8 @@ from nautilus_trader.backtest.engine import BacktestEngine
 from nautilus_trader.backtest.engine import BacktestEngineConfig
 from nautilus_trader.backtest.results import BacktestResult
 from nautilus_trader.common.actor import Actor
-from nautilus_trader.common.clock import LiveClock
 from nautilus_trader.common.config import ActorFactory
 from nautilus_trader.common.config import ImportableActorConfig
-from nautilus_trader.common.logging import Logger
-from nautilus_trader.common.logging import LoggerAdapter
-from nautilus_trader.common.logging import LogLevel
 from nautilus_trader.core.inspect import is_nautilus_class
 from nautilus_trader.model.currency import Currency
 from nautilus_trader.model.data.bar import Bar
@@ -68,7 +56,6 @@ from nautilus_trader.persistence.config import PersistenceConfig
 from nautilus_trader.persistence.streaming import FeatherWriter
 from nautilus_trader.trading.config import ImportableStrategyConfig
 from nautilus_trader.trading.config import StrategyFactory
-from nautilus_trader.trading.config import TradingStrategyConfig
 from nautilus_trader.trading.strategy import TradingStrategy
 
 
@@ -111,126 +98,6 @@ class BacktestNode:
             results.append(result)
 
         return self._gather_delayed(results)
-
-    def set_strategy_config(
-        self,
-        path: str,
-        strategy: TradingStrategyConfig,
-        instrument_id: str,
-        bar_type: str,
-        trade_size: Decimal,
-    ) -> None:
-        """
-        Set strategy parameters which can be passed to the hyperopt objective.
-
-        Parameters
-        ----------
-        path : str
-            The path to the strategy.
-        strategy : TradingStrategyConfig
-            The strategy config object.
-        instrument_id : InstrumentId
-            The instrument ID.
-        bar_type : BarType
-            The type of bar type used.
-        trade_size : Decimal
-            The trade size to be used.
-
-        """
-        self.path = path
-        self.strategy = strategy
-        self.instrument_id = instrument_id
-        self.bar_type = bar_type
-        self.trade_size = trade_size
-
-    def hyperopt_search(self, config, params, max_evals=50) -> Dict:
-        """
-        Run hyperopt to optimize strategy parameters.
-
-        Parameters
-        ----------
-        config : BacktestRunConfig
-            The configuration for the backtest test.
-        params : Dict[str, Any]
-            The set of strategy parameters to optimize.
-        max_evals : int
-            The maximum number of evaluations for the optimization problem.
-
-        Returns
-        -------
-        Dict
-            The optimized strategy parameters.
-
-        Raises
-        ------
-        ImportError
-            If hyperopt is not available.
-
-        """
-        if hyperopt is None:
-            raise ImportError(
-                "The hyperopt package is not installed. "
-                "Please install via pip or poetry install -E hyperopt",
-            )
-
-        logger = Logger(clock=LiveClock(), level_stdout=LogLevel.INFO)
-        logger_adapter = LoggerAdapter(component_name="HYPEROPT_LOGGER", logger=logger)
-        self.config = config
-
-        def objective(args):
-
-            logger_adapter.info(f"{args}")
-
-            strategies = [
-                ImportableStrategyConfig(
-                    path=self.path,
-                    config=self.strategy(
-                        instrument_id=self.instrument_id,
-                        bar_type=self.bar_type,
-                        trade_size=self.trade_size,
-                        **args,
-                    ),
-                ),
-            ]
-
-            local_config = self.config
-            local_config = local_config.replace(strategies=strategies)
-
-            local_config.check()
-
-            try:
-                result = self._run(
-                    engine_config=local_config.engine,
-                    run_config_id=local_config.id,
-                    venue_configs=local_config.venues,
-                    data_configs=local_config.data,
-                    actor_configs=local_config.actors,
-                    strategy_configs=local_config.strategies,
-                    persistence=local_config.persistence,
-                    batch_size_bytes=local_config.batch_size_bytes,
-                    # return_engine=True
-                )
-
-                base_currency = self.config.venues[0].base_currency
-                # logger_adapter.info(f"{result.stats_pnls[base_currency]}")
-                pnl_pct = result.stats_pnls[base_currency]["PnL%"]
-                logger_adapter.info(f"OBJECTIVE: {1/pnl_pct}")
-
-                if (1 / pnl_pct) == 0 or pnl_pct <= 0:
-                    ret = {"status": hyperopt.STATUS_FAIL}
-                else:
-                    ret = {"status": hyperopt.STATUS_OK, "loss": (1 / pnl_pct)}
-
-            except Exception as e:
-                ret = {"status": hyperopt.STATUS_FAIL}
-                logger_adapter.error(f"Bankruptcy : {e} ")
-            return ret
-
-        trials = hyperopt.Trials()
-
-        return hyperopt.fmin(
-            objective, params, algo=hyperopt.tpe.suggest, trials=trials, max_evals=max_evals
-        )
 
     def run_sync(self, run_configs: List[BacktestRunConfig], **kwargs) -> List[BacktestResult]:
         """
