+++ nautilus_trader/backtest/results.py
@@ -60,9 +60,9 @@ class BacktestResult:
 
 def ensure_plotting(func):
     """
-    Decorate a function that require a plotting library
+    Decorate a function that require a plotting library.
 
-    Ensures library is installed and providers a better error about how to install if not found
+    Ensures library is installed and providers a better error about how to install if not found.
     """
 
     def inner(*args, **kwargs):
